Refactoring; added exceptions; reworked dependencies.

+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Fee.cs
@@ -6,10 +6,10 @@
         public string Name { get; set; }
         public int Quantity { get; set; }
         public string Unit { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
         public int TaxRate { get; set; }
-        public decimal TaxAmount { get; set; }
-        public decimal GrossTotalAmount { get; set; }
-        public decimal NetTotalAmount { get; set; }
+        public int TaxAmount { get; set; }
+        public int GrossTotalAmount { get; set; }
+        public int NetTotalAmount { get; set; }
     }
 }
